Restruct Ptrs and Cleaned Up Warnings
Big Change, since parameters are stores in the action card tuples and we have access to the current AID already no need for arguments in the methods handling the actions. This eliminates all of our warnings for unused parameters.

Shuffle tableau Methods consoldiated into ShuffleTableauPlayed. Thsi can shuffle any amount of cards in deck

closet crew and closet blue have been merged into closetOfColorToFront.

endOfTurn has defualt arguments that allow for altering how end of turn happens. Helps with drawing extra actions, shuffle at end of turn, and skip player.

After you was simplified by just changing the current player index to easy alter another player with out really changing whos turn it is.

Add to tableau was generalized. So add three and add 1  have been merged

meryl to front was replaced with closetIDToFront. This allows other spefic id's to be moved to front.

+++ foOscars/gamestate.cpp
@@ -86,7 +86,7 @@ QByteArray GameState::serialize()
     QJsonDocument jsonDoc;
     QJsonArray jsonPlayerArr;
 
-    for(int i =0; i<players.size(); i++){
+    for(int i =0; i<(int)players.size(); i++){
         Player player = players.at(i);
 
         QJsonObject playerObj;
